Add concept_id query
