+++ stac-gen/requirements.txt
@@ -1,0 +1,2 @@
+python-cmr
+awslambdaric
