--- stac/handler.py
@@ -1,12 +1,0 @@
-def handler(event, context):
-    """
-    Lambda handler for STAC Collection Item generation
-    """
-
-
-if __name__ == "__main__":
-    sample_event = {
-
-    }
-
-    handler(sample_event, {})
